* Include i386 architecture in the static library.
The resulting library will be placed in $(BUILT_PRODUCTS_DIR)/Debug-universal or Release-universal.

The script from the following stack overflow answer has been used: http://stackoverflow.com/a/3647187/346029. The strategy is to compile the _other_ architecture via xcbuild, then combine them both with lipo.

In addition, the path for the public headers has been set to /include (which is relative to the installation directory) so both a "copy headers" phase and the script work. In fact, a "copy headers" phase is now used instead of the "copy files" phase that was used for the same purpose.